recommendation/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
pyarrow==15.0.2
streamlit==1.33.0
torch==2.2.2
plotly==5.22.0
google-cloud-bigquery==3.25.0
db-dtypes==1.2.0
